separated App and StateMachine and changed useTimer arguments
useTimer now accepts one object instead of using overloads and it
includes the option to add dependencies to the internal timer hook so
that it gets reset when one of the external dependencies changes.
