+++ src/components/BoardText.tsx
@@ -1,4 +1,4 @@
-import { StateMachineSlideProps, useTimer } from '../App';
+import { StateMachineSlideProps, useTimer } from '../StateMachine';
 import { useBoardMessagesQuery } from '../store/api';
 
 export default function BoardText({ current }: StateMachineSlideProps) {
