+++ src/components/Quotes.tsx
@@ -1,4 +1,4 @@
-import { StateMachineSlideProps, useTimer } from '../App';
+import { StateMachineSlideProps, useTimer } from '../StateMachine';
 import { useQuotesQuery } from '../store/api';
 
 export default function Quotes({ current }: StateMachineSlideProps) {
